requirements.txt
pytest>=9.0.0
PyYAML>=6.0
